automatedimageprocess/requirements.txt
numpy
opencv-python
matplotlib
scipy
streamlit
Pillow
